+++ GEMA/__init__.py
@@ -2,6 +2,7 @@ from .map import Map
 from .classification import Classification
 from .visualization import Visualization
 from .iterativesom import IterativeSOM
+from .accelerator import CUPY_AVAILABLE, NUMBA_AVAILABLE, device_info
 
 name = "GEMA"
-__version__ = "0.5.1"
+__version__ = "0.5.2"